Fix #25 by providing an Akka Stream Source graph stage for Kinesis. (#39)
Adds an Akka Streams `Source` is that can be used with streams. It is possible to create a source from a ConsumerConf or directly from the consumer name that is defined in the configuration. Every message that is emitted to the stream is of type CommitableEvent[ConsumerEvent] and has to be committed explicitly downstream with a call to event.commit(). It is possible to map to a different type of CommittableEvent via the map and mapAsync functionality.

A KinesisConsumer is used internally for the Kinesis.source. All rules described here for the KinesisConsumer also apply for the stream source.